Add Firebase integration
